List branches regardless of cwd
I found that the branch and remote completion was only working if your cwd was where the .git folder existed. I added a recursive method that will keep traversing up the path until it finds a .get folder. It's kind of ugly, but I suck at lua. Hopefully you can either accept the PR, or this will inspire you to write it better!