Hack to stop recursion
Sorry but this is a hack. But there had to be a way to stop recursing if no .get folder is ever found. Otherwise you keep running chdir to the root of the drive. And I've even hard coded the C drive. Really sorry. Hopefully you can shed some light on a better way.