Add support for working in directories created with git worktree add
git worktree add makes directories that do not have a .git directory.
Instead, .git is a text file that looks like this:

    :~/gittest/myworktree$ cat .git
    gitdir: /home/user/gittest/dirContainingGitRepo/.git/worktrees/myworktree

This allows the user to have multiple worktrees that have builds of
different branches without having to duplicate the .git directory for
all of them.

In this case, the existing lua code doesn't work, as it is looking in
the wrong place for the list of branches.