`git stash ...` works with `clink-popup-complete`.
In Clink v1.1.12 and higher `clink.ondisplaymatches` lets a match
generator supply descriptions associated with matches.  This makes it
possible for the `clink-popup-complete` command to show stash
descriptions next to the stash IDs.

As with the preceding change, both of these are written to be
compatible with all versions of Clink.