Restructure code to satisfy code coverage.
It uses the short circuiting `and` operator to try to trick code
coverage analysis into considering the line to have been executed even
though it's impossible to execute.

I could invest in building a mock infrastructure for this repo, but
that's more cost than I'm willing to invest for this scenario.