[git] Colored completions for `git checkout ...` with Clink v1.1.12 popup list (#135)
* [git] Optionally color the `*` in checkout completions.

In Clink v1.1.9 and higher a match display filter can provide colored
text.  The `git checkout` match display filter now applies a color to
the `*` next to preferred branch names.

This change is written to be compatible with all versions of Clink.

* [git] `git stash ...` works with `clink-popup-complete`.

In Clink v1.1.12 and higher `clink.ondisplaymatches` lets a match
generator supply descriptions associated with matches.  This makes it
possible for the `clink-popup-complete` command to show stash
descriptions next to the stash IDs.

This change is written to be compatible with all versions of Clink.

* Add a version module and color helper.

* Busted runs the modules scripts outside of Clink.

The version module is meant to encapsulate Clink versioning, but it must
do so without relying on Clink being present.

* Restructure code to satisfy code coverage.

It uses the short circuiting `and` operator to try to trick code
coverage analysis into considering the line to have been executed even
though it's impossible to execute.

I could invest in building a mock infrastructure for this repo, but
that's more cost than I'm willing to invest for this scenario.

* Address additional PR feedback:
- Rename the version module.
- Compute all version checks during script exec and cache the results.
- Add comment about Clink's color setting format being SGR params.

* Exclude the renamed version module from codecov.

* Address additional PR feedback:
- Factor the colored star construction differently.

* Fix typo.
Somehow I renamed version module name in the require(), but accidentally
undid the change in this copy of the file.