Fix #117; .init.lua isn't loaded under some circumstances.
The "." at the beginning of the ".init.lua" file name marks it as a hidden file
in Linux and some sync programs.  That can cause sync programs to not copy the
file.  Renaming to "!init.lua" resolves that problem, while still ensuring the
script is loaded before any of the other completion scripts.