Perf improvement in edge case.
The git_prompt.lua prompt filter tries to add " -> remote" to the branch
name in the prompt, and it loads the git config file to do so.

In some repos, the git config file may be large enough that it could
cause a perceivable delay.  Using a prompt coroutine enables caching the
result so that the config file is only loaded once per input prompt,
even if other async prompt filters trigger multiple refreshes.

Tested with Clink 1.5.1:
- Outside a git repo.
- Inside a git repo with a remote.
- Inside a git repo without a remote.

Tested with Clink 1.1.12 (does not support prompt coroutines):
- Outside a git repo.
- Inside a git repo with a remote.
- Inside a git repo without a remote.