Always complete local files for `git log` and `git diff`.
Commit 321eb3a9330eab1948011f9f25ec9a7fd5c84302 addressed #199 to make
completions for `git log` and `git diff` (and `git checkout`) include
tags, branches, and remotes.

In the source code for bash's git completions, it clearly did not
generate file matches unless either `git log ./` or `git log --` were
used.  And empirically, that seemed to accurately match the behavior.

But I missed that the way completions work in `bash`, if completion
scripts produce _zero_ matches then `bash` falls back to local file
completion.  So the change had the effect of making the list of
possible completions match what is reported in bash ... but accidentally
made the commands unable to complete local files without using either
`./` or `--`.

This new commit changes the `git log` and `git diff` (and
`git checkout`) completions to always include local files.  The list of
possible completions is larger than in `bash`, but it's more complete
and accurate, and enables other enhancements in Clink to work on the
whole list of possible completions, all the time.